Move peer-pods code to its own file
$ wc -l controllers/openshift_controller.go
2764 controllers/openshift_controller.go

The file is huge already.

The peer-pods logic needs some heavy fixing that will require to add some
more peer-pods specific code.

Let's move peer-pods to a separate file. It will be easier to read and
maintain. This is consistent with what we already do for other more recent
features.

This doesn't change behavior.
