+++ controllers/peerpods.go
@@ -1,0 +1,506 @@
+/*
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package controllers
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"path/filepath"
+	"time"
+
+	mcfgv1 "github.com/openshift/api/machineconfiguration/v1"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+const (
+	DEFAULT_PEER_PODS                   = "10"
+	peerpodConfigCrdName                = "peerpodconfig-openshift"
+	peerpodsMachineConfigPathLocation   = "/config/peerpods"
+	peerpodsCrioMachineConfig           = "50-kata-remote"
+	peerpodsCrioMachineConfigYaml       = "mc-50-crio-config.yaml"
+	peerpodsKataRemoteMachineConfig     = "40-worker-kata-remote-config"
+	peerpodsKataRemoteMachineConfigYaml = "mc-40-kata-remote-config.yaml"
+	// https://github.com/kata-containers/kata-containers/blob/main/tools/packaging/kata-deploy/runtimeclasses/kata-remote.yaml#L7
+	peerpodsRuntimeClassName        = "kata-remote"
+	peerpodsRuntimeClassCpuOverhead = "0.25"
+	peerpodsRuntimeClassMemOverhead = "120Mi"
+	// cloud-api-adaptor (CAA) daemonset name
+	caaDsName = "osc-caa-ds"
+)
+
+func MountProgagationRef(mode corev1.MountPropagationMode) *corev1.MountPropagationMode {
+	return &mode
+}
+
+func (r *KataConfigOpenShiftReconciler) processDaemonsetForCAA() *appsv1.DaemonSet {
+	var (
+		runPrivileged                = true
+		runAsUser              int64 = 0
+		defaultMode            int32 = 0600
+		sshSecretOptional            = true
+		authJsonSecretOptional       = true
+		nodeSelector                 = r.getNodeSelectorAsMap()
+	)
+
+	dsLabelSelectors := map[string]string{
+		"name": caaDsName,
+	}
+
+	imageString := os.Getenv("RELATED_IMAGE_CAA")
+	if imageString == "" {
+		r.Log.Info("RELATED_IMAGE_CAA env var is unset or empty, cloud-api-adaptor pods will not run")
+	}
+
+	return &appsv1.DaemonSet{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "apps/v1",
+			Kind:       "DaemonSet",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      caaDsName,
+			Namespace: os.Getenv("PEERPODS_NAMESPACE"),
+		},
+		Spec: appsv1.DaemonSetSpec{
+			Selector: &metav1.LabelSelector{
+				MatchLabels: dsLabelSelectors,
+			},
+			UpdateStrategy: appsv1.DaemonSetUpdateStrategy{
+				Type: "RollingUpdate",
+				RollingUpdate: &appsv1.RollingUpdateDaemonSet{
+					MaxUnavailable: &intstr.IntOrString{
+						Type:   intstr.Int,
+						IntVal: 1,
+					},
+				},
+			},
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels: dsLabelSelectors,
+				},
+				Spec: corev1.PodSpec{
+					ServiceAccountName: "default",
+					NodeSelector:       nodeSelector,
+					HostNetwork:        true,
+					Containers: []corev1.Container{
+						{
+							Name:            "caa-pod",
+							Image:           imageString,
+							ImagePullPolicy: "Always",
+							SecurityContext: &corev1.SecurityContext{
+								// TODO - do we really need to run as root?
+								Privileged: &runPrivileged,
+								RunAsUser:  &runAsUser,
+							},
+							Command: []string{"/usr/local/bin/entrypoint.sh"},
+							Env: []corev1.EnvVar{
+								{
+									Name: "NODE_NAME",
+									ValueFrom: &corev1.EnvVarSource{
+										FieldRef: &corev1.ObjectFieldSelector{
+											FieldPath: "spec.nodeName",
+										},
+									},
+								},
+							},
+							EnvFrom: []corev1.EnvFromSource{
+								{
+									SecretRef: &corev1.SecretEnvSource{
+										LocalObjectReference: corev1.LocalObjectReference{
+											Name: "peer-pods-secret",
+										},
+									},
+								},
+								{
+									ConfigMapRef: &corev1.ConfigMapEnvSource{
+										LocalObjectReference: corev1.LocalObjectReference{
+											Name: "peer-pods-cm",
+										},
+									},
+								},
+							},
+							VolumeMounts: []corev1.VolumeMount{
+								{
+									Name:      "auth-json-volume",
+									MountPath: "/root/containers/",
+									ReadOnly:  true,
+								}, {
+									Name:      "ssh",
+									MountPath: "/root/.ssh",
+									ReadOnly:  true,
+								},
+								{
+									MountPath: "/run/peerpod",
+									Name:      "pods-dir",
+								},
+								{
+									MountPath:        "/run/netns",
+									MountPropagation: MountProgagationRef(corev1.MountPropagationHostToContainer),
+									Name:             "netns",
+								},
+							},
+						},
+					},
+					Volumes: []corev1.Volume{
+						{
+							Name: "auth-json-volume",
+							VolumeSource: corev1.VolumeSource{
+								Secret: &corev1.SecretVolumeSource{
+									SecretName:  "auth-json-secret",
+									DefaultMode: &defaultMode,
+									Optional:    &authJsonSecretOptional,
+								},
+							},
+						}, {
+							Name: "ssh",
+							VolumeSource: corev1.VolumeSource{
+								Secret: &corev1.SecretVolumeSource{
+									SecretName:  "ssh-key-secret",
+									DefaultMode: &defaultMode,
+									Optional:    &sshSecretOptional,
+								},
+							},
+						},
+						{
+							Name: "pods-dir",
+							VolumeSource: corev1.VolumeSource{
+								HostPath: &corev1.HostPathVolumeSource{
+									Path: "/run/peerpod",
+								},
+							},
+						},
+						{
+							Name: "netns",
+							VolumeSource: corev1.VolumeSource{
+								HostPath: &corev1.HostPathVolumeSource{
+									Path: "/run/netns",
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+// Handles provider specific parts of the CAA Ds
+// Modifies the DaemonSet if needed
+func (r *KataConfigOpenShiftReconciler) processProviderConfigCAA(ds *appsv1.DaemonSet) error {
+	r.Log.Info("Getting cloud provider from infra")
+	provider, err := getCloudProviderFromInfra(r.Client)
+	if err != nil {
+		return fmt.Errorf("failed to get cloud provider from infra: %w", err)
+	}
+
+	switch provider {
+	case IBMCloudProvider:
+		var expSecs int64 = 3600
+		vaultTokenVolume := corev1.Volume{
+			Name: "vault-token",
+			VolumeSource: corev1.VolumeSource{
+				Projected: &corev1.ProjectedVolumeSource{
+					Sources: []corev1.VolumeProjection{
+						{
+							ServiceAccountToken: &corev1.ServiceAccountTokenProjection{
+								Path:              "vault-token",
+								ExpirationSeconds: &expSecs,
+								Audience:          "iam",
+							},
+						},
+					},
+				},
+			},
+		}
+
+		vaultTokenVolumeMount := corev1.VolumeMount{
+			MountPath: "/var/run/secrets/tokens",
+			Name:      "vault-token",
+		}
+
+		ds.Spec.Template.Spec.Volumes = append(ds.Spec.Template.Spec.Volumes, vaultTokenVolume)
+		for i := range ds.Spec.Template.Spec.Containers {
+			container := &ds.Spec.Template.Spec.Containers[i]
+			if container.Name == "caa-pod" {
+				container.VolumeMounts = append(container.VolumeMounts, vaultTokenVolumeMount)
+			}
+		}
+
+		return nil
+	default:
+		return nil
+	}
+}
+
+func (r *KataConfigOpenShiftReconciler) createAuthJsonSecret() error {
+	var err error
+
+	pullSecret := &corev1.Secret{}
+	err = r.Client.Get(context.TODO(), types.NamespacedName{Name: "pull-secret", Namespace: "openshift-config"}, pullSecret)
+	if err != nil {
+		r.Log.Info("Error fetching pull-secret", "err", err)
+		return err
+	}
+
+	authJsonSecret := corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "auth-json-secret",
+			Namespace: OperatorNamespace,
+		},
+		Data: map[string][]byte{
+			"auth.json": pullSecret.Data[".dockerconfigjson"],
+		},
+		Type: corev1.SecretTypeOpaque,
+	}
+
+	err = r.Client.Create(context.TODO(), &authJsonSecret)
+	if err != nil {
+		if k8serrors.IsAlreadyExists(err) {
+			err = r.Client.Update(context.TODO(), &authJsonSecret)
+			if err != nil {
+				r.Log.Info("Error updating auth-json-secret", "err", err)
+				return err
+			}
+		} else {
+			r.Log.Info("Error creating auth-json-secret", "err", err)
+			return err
+		}
+	}
+
+	return err
+}
+
+// Create the MachineConfigs for PeerPod
+// We do it before kata-oc creation to optimise the reboots required for MC creation
+func (r *KataConfigOpenShiftReconciler) enablePeerPodsMc() error {
+
+	//Create MachineConfig for kata-remote hyp CRIO config
+	crioMachineConfigFilePath := filepath.Join(peerpodsMachineConfigPathLocation, peerpodsCrioMachineConfigYaml)
+	err := r.createMcFromFile(crioMachineConfigFilePath)
+	if err != nil {
+		r.Log.Info("Error in creating CRIO MachineConfig", "err", err)
+		return err
+	}
+
+	//Create MachineConfig for kata-remote hyp config toml
+	kataConfigMachineConfigFilePath := filepath.Join(peerpodsMachineConfigPathLocation, peerpodsKataRemoteMachineConfigYaml)
+	err = r.createMcFromFile(kataConfigMachineConfigFilePath)
+	if err != nil {
+		r.Log.Info("Error in creating kata remote configuration.toml MachineConfig", "err", err)
+		return err
+	}
+
+	return nil
+}
+
+// Create the PeerPodConfig CRDs and misc configs required for peer-pods
+func (r *KataConfigOpenShiftReconciler) enablePeerPodsMiscConfigs() error {
+	// Create the CAA daemonset
+	ds := r.processDaemonsetForCAA()
+	if err := r.processProviderConfigCAA(ds); err != nil {
+		r.Log.Error(err, "Failed setting cloud provider specific configuration for cloud-api-adaptor DS")
+		return err
+	}
+	r.Log.Info("Got CAA ds manifest", "ds", ds)
+
+	if err := controllerutil.SetControllerReference(r.kataConfig, ds, r.Scheme); err != nil {
+		r.Log.Error(err, "Failed setting ControllerReference for cloud-api-adaptor DS")
+		return err
+	}
+
+	err := r.Client.Update(context.TODO(), ds)
+	if err != nil && k8serrors.IsNotFound(err) {
+		r.Log.Error(err, "cloud-api-adaptor daemonset doesn't exist. Creating")
+		err = r.Client.Create(context.TODO(), ds)
+		if err != nil {
+			r.Log.Error(err, "failed to create cloud-api-adaptor daemonset")
+			return err
+		}
+	}
+
+	// Create the mutating webhook deployment
+	err = r.createMutatingWebhookDeployment()
+	if err != nil {
+		r.Log.Info("Error in creating mutating webhook deployment for peerpods", "err", err)
+		return err
+	}
+
+	// Create the mutating webhook service
+	err = r.createMutatingWebhookService()
+	if err != nil {
+		r.Log.Info("Error in creating mutating webhook service for peerpods", "err", err)
+		return err
+	}
+
+	// Create the mutating webhook
+	err = r.createMutatingWebhookConfig()
+	if err != nil {
+		r.Log.Info("Error in creating mutating webhook for peerpods", "err", err)
+		return err
+	}
+
+	// Create runtimeClass config for peer-pods
+	err = r.createRuntimeClass(peerpodsRuntimeClassName, peerpodsRuntimeClassCpuOverhead, peerpodsRuntimeClassMemOverhead)
+	if err != nil {
+		r.Log.Info("Error in creating kata remote runtimeclass", "err", err)
+		return err
+	}
+	return nil
+}
+
+func (r *KataConfigOpenShiftReconciler) disablePeerPods() error {
+	ds := r.processDaemonsetForCAA()
+	err := r.Client.Delete(context.TODO(), ds)
+	if err != nil {
+		if k8serrors.IsNotFound(err) {
+			r.Log.Info("cloud-api-adaptor daemonset was already deleted")
+		} else {
+			r.Log.Error(err, "error when deleting cloud-api-adaptor Daemonset, try again")
+			return err
+		}
+	}
+
+	if r.DeploymentMode == MachineConfigMode {
+		// We are explicitly ignoring any errors in peerpodconfig and related machineconfigs removal as
+		// these can be removed manually if needed and this is not in the critical path
+		// of operator functionality
+		_ = r.deletePeerPodsMC()
+	}
+
+	if r.DeploymentMode == DaemonSetMode {
+		// We want to make sure that the osc-config-sync ds removal successfully started
+		// So we will try again in case of an error
+		err = r.deletePeedPodsConfigDaemonSet()
+		if err != nil {
+			return err
+		}
+	}
+
+	// Delete mutating webhook deployment
+	err = r.deleteMutatingWebhookDeployment()
+	if err != nil {
+		r.Log.Info("Error in deleting mutating webhook deployment for peerpods", "err", err)
+		return err
+	}
+
+	// Delete mutating webhook service
+	err = r.deleteMutatingWebhookService()
+	if err != nil {
+		r.Log.Info("Error in deleting mutating webhook service for peerpods", "err", err)
+		return err
+	}
+
+	// Delete the mutating webhook
+	err = r.deleteMutatingWebhookConfig()
+	if err != nil {
+		r.Log.Info("Error in deleting mutating webhook for peerpods", "err", err)
+		return err
+	}
+
+	return nil
+}
+
+func (r *KataConfigOpenShiftReconciler) deletePodVMImage() (*ctrl.Result, error) {
+	// Handle podvm image deletion
+	// Since we want to declaratively reach the final state, we need to reconcile when there are errors
+	// as we want the system to give a chance of fixing the error.
+	// For cases we don't want to reconcile, ie for ImageDeletedSuccessfully and UnsupportedPodVMImageProvider
+	// we should just log the message and let the code continue without explicitly returning from the method
+
+	// Following are returned statuses:
+	// ImageDeletedSuccessfully
+	// UnsupportedPodVMImageProvider
+	// ImageDeletionFailed
+	// RequeueNeeded
+	// ImageDeletionStatusUnknown
+	status, err := ImageDelete(r.Client)
+	switch status {
+	case ImageDeletedSuccessfully:
+		r.setInProgressConditionToPodVMImageDeleted()
+		r.Log.Info("PodVM Image deleted successfully")
+
+	case UnsupportedPodVMImageProvider:
+		r.setInProgressConditionToPodVMImageUnsupportedProvider()
+		r.Log.Info("unsupported cloud provider, skipping image deletion")
+
+	case RequeueNeeded:
+		r.setInProgressConditionToPodVMImageDeleting()
+		return &ctrl.Result{Requeue: true, RequeueAfter: 15 * time.Second}, err
+
+	case ImageDeletionFailed:
+		r.setInProgressConditionToPodVMImageDeletionFailed()
+		if err != nil {
+			// We requeue only if there is an error.
+			return &ctrl.Result{Requeue: true, RequeueAfter: 15 * time.Second}, err
+		}
+		// If there's no error, log and continue
+		r.Log.Info("Image deletion failed. Check logs for more details")
+
+	case ImageDeletionStatusUnknown:
+		r.setInProgressConditionToPodVMImageDeletionUnknown()
+		return &ctrl.Result{Requeue: true, RequeueAfter: 15 * time.Second}, err
+
+	default:
+		// For all other statuses, just log and continue
+		r.Log.Info("PodVM Image deletion status and error", "status", status, "error", err)
+	}
+	return nil, nil
+}
+
+func (r *KataConfigOpenShiftReconciler) deletePeerPodsMC() error {
+	mc := mcfgv1.MachineConfig{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "machineconfiguration.openshift.io/v1",
+			Kind:       "MachineConfig",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: peerpodsKataRemoteMachineConfig,
+		},
+	}
+
+	err := r.Client.Delete(context.TODO(), &mc)
+	if err != nil {
+		// error during removing mc. Just log the error and move on.
+		r.Log.Info("Error found deleting mc. If the MachineConfig object exists after uninstallation it can be safely deleted manually",
+			"mc", mc.Name, "err", err)
+	}
+
+	mc = mcfgv1.MachineConfig{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "machineconfiguration.openshift.io/v1",
+			Kind:       "MachineConfig",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: peerpodsCrioMachineConfig,
+		},
+	}
+
+	err = r.Client.Delete(context.TODO(), &mc)
+	if err != nil {
+		// error during removing mc. Just log the error and move on.
+		r.Log.Info("Error found deleting mc. If the MachineConfig object exists after uninstallation it can be safely deleted manually",
+			"mc", mc.Name, "err", err)
+	}
+
+	return nil
+}
